refactor: Timeline構造を修正
