+++ frontend/src/components/timeline/formatTime.ts
@@ -1,0 +1,7 @@
+export const formatTime = (date: Date): string => {
+  return date.toLocaleTimeString('ja-JP', {
+    hour: '2-digit',
+    minute: '2-digit',
+    hour12: false,
+  });
+};
